Move vanilla react code to next app
