+++ frontend/pages/index.tsx
@@ -1,0 +1,29 @@
+import type { NextPage } from 'next';
+import { ReactElement } from 'react';
+import styled from 'styled-components';
+import { ActivateDeactivate } from '../components/ActivateDeactivate';
+import { Greeter } from '../components/Greeter';
+import { SectionDivider } from '../components/SectionDivider';
+import { SignMessage } from '../components/SignMessage';
+import { WalletStatus } from '../components/WalletStatus';
+
+const StyledAppDiv = styled.div`
+  display: grid;
+  grid-gap: 20px;
+`;
+
+const Home: NextPage = () => {
+  return (
+    <StyledAppDiv>
+      <ActivateDeactivate />
+      <SectionDivider />
+      <WalletStatus />
+      <SectionDivider />
+      <SignMessage />
+      <SectionDivider />
+      <Greeter />
+    </StyledAppDiv>
+  )
+}
+
+export default Home
