--- frontend/public/manifest.json
@@ -1,9 +1,0 @@
-{
-  "short_name": "Ethereum Hardhat Typescript Starter Kit",
-  "name": "Ethereum Hardhat Typescript Starter Kit",
-  "start_url": ".",
-  "display": "standalone",
-  "theme_color": "#000000",
-  "background_color": "#ffffff"
-}
-
