--- frontend/src/index.tsx
@@ -1,15 +1,0 @@
-import { Web3ReactProvider } from '@web3-react/core';
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { App } from './App';
-import './index.css';
-import { getProvider } from './utils/provider';
-
-ReactDOM.render(
-  <React.StrictMode>
-    <Web3ReactProvider getLibrary={getProvider}>
-      <App />
-    </Web3ReactProvider>
-  </React.StrictMode>,
-  document.getElementById('root')
-);
