--- frontend2/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "frontend2",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "next": "12.1.6",
-    "react": "18.1.0",
-    "react-dom": "18.1.0"
-  },
-  "devDependencies": {
-    "@types/node": "17.0.31",
-    "@types/react": "18.0.9",
-    "@types/react-dom": "18.0.3",
-    "eslint": "8.15.0",
-    "eslint-config-next": "12.1.6",
-    "typescript": "4.6.4"
-  }
-}
